Tweaked feedback routing and removed EMI filter
The EMI filter places an inductor in between the battery and the buck
converter, creating the risk of resonance either at the buck converter
input or on the battery line. While the former issue has been *somewhat*
addressed with the parallel damping resistor, the latter issue cannot be
reasonably fixed without good knowledge of the other various loads that
may be attached to the battery line. For this reason, the EMI filter is
removed.